+++ CHANGELOG.md
@@ -11,6 +11,7 @@ against this Rust version.
 - [added] Playground metadata
 - [changed] Remove build scripts; instead check-in the built code
 - [added] Constants for 3600
+- [added] `FoldAdd` and `FoldMul` to get the sum/product of an array
 
 ### 1.16.0 (2022-12-05)
 - [added] `const INT` field to the `ToInt` trait.